refactor(doctor): break the escalation import cycles
check:cycles caught three new cycles, all rooted in one edge:
doctor-repair imported doctor.ts for runDetailedPluginHealthChecks, so
the new doctor → doctor-escalation → (repair stack) imports closed loops.

Two moves, no behavior change:
- The plugin health-check RUNNER moves to a leaf module
  (doctor-checks.ts); doctor.ts re-exports it for existing consumers and
  doctor-repair imports the leaf directly.
- notifyUnfixableIssues + its dedup cache move into doctor-escalation.ts
  (they are escalation behavior); doctor.ts imports them one-way.

check:cycles: 0 new. Full suite green.
